+++ src/api/index.ts
@@ -1,4 +1,3 @@
-
 type TRequest = {
   url: string;
   init?: RequestInit;
@@ -8,6 +7,6 @@ export class Fetcher {
   static baseApiUrl = `${window.location.protocol}//${window.location.hostname}:${process.env.REACT_PORT_API}`;
 
   public static async request({url, init}: TRequest) {
-    return await fetch(`${this.baseApiUrl}/${url}`, init);
+    return fetch(`${this.baseApiUrl}/${url}`, init);
   }
 }
